test(dbcontext): replace old dbcontext tests and add failure-reason handling coverage
