+++ packages/grzctl/tests/test_dbcontext.py
@@ -1,3 +1,6 @@
+from unittest import mock
+from unittest.mock import MagicMock
+
 import pytest
 from grz_common.exceptions import (
     DecryptionError,
@@ -7,7 +10,7 @@ from grz_common.exceptions import (
     UploadError,
 )
 from grz_db.errors import DuplicateTanGError
-from grz_db.models.submission import FailureReasonEnum
+from grz_db.models.submission import FailureReasonEnum, SubmissionStateEnum
 from grzctl.dbcontext import DbContext
 from pydantic import ValidationError
 
@@ -19,6 +22,24 @@ def db_context() -> DbContext:
     ctx.enabled = False
     ctx.db = None
     return ctx
+
+
+@pytest.fixture
+def mock_db():
+    return MagicMock()
+
+
+@pytest.fixture
+def ctx(mock_db):
+    context = DbContext(
+        configuration={},
+        submission_id="123_2025-01-01_00000000",
+        start_state=SubmissionStateEnum.ENCRYPTING,
+        end_state=SubmissionStateEnum.ENCRYPTED,
+        enabled=True,
+    )
+    context.db = mock_db  # bypass __enter__
+    return context
 
 
 class TestMapExceptionToFailureReason:
@@ -95,3 +116,46 @@ class TestMapExceptionToFailureReason:
         }
         unmapped = {e for e in FailureReasonEnum if e != FailureReasonEnum.UNKNOWN} - mapped_results
         assert not unmapped, f"These FailureReasonEnum values have no exception mapping: {unmapped}"
+
+
+class TestDbContextFailureReason:
+    def test_file_not_found_maps_correctly(self, ctx, mock_db):
+        exc = FileNotFoundError("missing file")
+        ctx.__exit__(type(exc), exc, None)
+        mock_db.update_submission_state.assert_called_once_with(
+            ctx.submission_id,
+            SubmissionStateEnum.ERROR,
+            data={"error": str(exc)},
+            failure_reason=FailureReasonEnum.FILE_NOT_FOUND,
+            grzctl_versions=mock.ANY,
+        )
+
+    def test_validation_error_maps_correctly(self, ctx, mock_db):
+        exc = ValidationError.from_exception_data("test", [])
+        ctx.__exit__(type(exc), exc, None)
+        mock_db.update_submission_state.assert_called_once_with(
+            ctx.submission_id,
+            SubmissionStateEnum.ERROR,
+            data={"error": str(exc)},
+            failure_reason=FailureReasonEnum.VALIDATION_ERROR,
+            grzctl_versions=mock.ANY,
+        )
+
+    def test_unknown_exception_maps_to_unknown(self, ctx, mock_db):
+        exc = RuntimeError("something unexpected")
+        ctx.__exit__(type(exc), exc, None)
+        mock_db.update_submission_state.assert_called_once_with(
+            ctx.submission_id,
+            SubmissionStateEnum.ERROR,
+            data={"error": str(exc)},
+            failure_reason=FailureReasonEnum.UNKNOWN,
+            grzctl_versions=mock.ANY,
+        )
+
+    def test_no_exception_does_not_set_failure_reason(self, ctx, mock_db):
+        ctx.__exit__(None, None, None)
+        mock_db.update_submission_state.assert_called_once_with(
+            ctx.submission_id,
+            SubmissionStateEnum.ENCRYPTED,
+            grzctl_versions=mock.ANY,
+        )
